refactor: rename api_v2 to api and remove file, use init instead
