+++ propelauth_py/api/user.py
@@ -1,9 +1,7 @@
 import requests
-import requests
-from requests.auth import AuthBase
 
 from propelauth_py.api import _ApiKeyAuth, _format_params, _is_valid_id
-from propelauth_py.api_v2.end_user_api_keys import _validate_api_key
+from propelauth_py.api.end_user_api_keys import _validate_api_key
 from propelauth_py.errors import (
     CreateUserException,
     EndUserApiKeyException,
